requirements.txt
# Google ADK (Agent Development Kit)
google-adk

# FastAPI and Uvicorn
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Additional dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# For timezone handling (used in aaggent)
tzdata>=2023.3

websockets
python-multipart
python-json-logger